Add tests for the elf dependency generator fallback version feature.
These tests are not compatible with fakechroot, for two reasons.
While fakechroot supports dlmopen(), it will convert relative paths
to absolute paths, which prevents library path searching.  The test
could be run from the data directory so that the relative path's
absolute path expansion was correct, and dlmopen() will succeed.
However, linkmap->l_name will have the real path, not a chrooted
path, and when that is passed to readlink(), fakechroot will try
to expand a path that already includes the chroot prefix.  So,
even when the library can be loaded with dlmopen(), the symlink
can't be resolved to find the path that contains the version.